% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_data_empirical.R
\docType{data}
\name{artificial_matlist}
\alias{artificial_matlist}
\alias{artificial_meta}
\alias{artificial_progress}
\alias{artificial_data}
\title{raw data for analysis of artificial data}
\format{
\code{artificial_matlist} is a large list with the individual
interaction matrices that were created for the removal experiment
(a total of 11,571 items, i.e., matrices)

\code{artificial_meta} is a large list (with 1,000 items
corresponding to 1,000 artificial data sets) with the settings
used to create the original data sets

\code{artificial_progress} is a data frame with 71,766 rows, which
was used to track progress in the analysis
}
\usage{
artificial_matlist
}
\description{
raw data for analysis of artificial data
}
\keyword{datasets}
