% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark_plot.R
\name{benchmark_plot}
\alias{benchmark_plot}
\title{benchmark plot}
\usage{
benchmark_plot(
  pdata,
  empirical = NULL,
  prunkcut = 0.05,
  benchmark = c("pij", "input")
)
}
\arguments{
\item{pdata}{\code{data("removal_global")}}

\item{empirical}{logical, should empirical data be plotted (if \code{TRUE}),
or artificial data (if \code{FALSE}). Default is \code{NULL}, which
combines both sets.}

\item{prunkcut}{numeric, the cutoff for unknown relationships. Default
is \code{0.05}.}

\item{benchmark}{character, either \code{"pij"} or \code{"input"}. With
\code{"pij"} steepness based on Pij David's score is used as ground
truth. With \code{"input"}, we use the steepness value that was used
during interaction data generation as ground truth, and this is only
applicable when \code{empirical = FALSE}.}
}
\value{
a plot and invisibly the results of correlation and
simple regression.
}
\description{
benchmark plot
}
\examples{
data("removal_global")
benchmark_plot(removal_global, empirical = TRUE, benchmark = "pij")
x <- benchmark_plot(removal_global, empirical = FALSE, benchmark = "input")
x
}
