% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_data_cred_width.R
\docType{data}
\name{cred_width}
\alias{cred_width}
\title{raw data for analysis of credible interval width as function of interaction ratios}
\format{
\code{cred_width} is a data frame with 1,670 observations, each line
corresponding to one matrix with steepness evaluated and extracted
width of the 89\\% credible interval
}
\usage{
cred_width
}
\description{
raw data for analysis of credible interval width as function of interaction ratios
}
\details{
The data sets correspond to the 1,000 artificial data sets we
created and the 670 empiricial data sets.
}
\keyword{datasets}
