% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_data_empirical_key.R
\docType{data}
\name{empirical_key}
\alias{empirical_key}
\title{empirical data summary and sources}
\format{
A \code{data.frame}.
}
\source{
Christof Neumann
}
\usage{
empirical_key
}
\description{
empirical data summary and sources
}
\keyword{datasets}
