% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_data_artificial.R
\docType{data}
\name{empirical_matlist}
\alias{empirical_matlist}
\alias{empirical_progress}
\alias{empirical_raw_data}
\alias{empirical_data}
\title{raw data for analysis of empirical data}
\format{
\code{empirical_matlist} is a large list with the individual
interaction matrices that were created for the removal experiment
(a total of 7,113 items, i.e., matrices)

\code{empirical_raw_data} is a large list with the individual
interaction matrices as they were taken from the literature. For
a full list of sources and references see \code{\link{empirical_key}}
and the vignette \code{vignette("empirical_references", package = "EloSteepness.data")}

\code{empirical_progress} is a data frame with 49,791 rows, which
was used to track progress in the analysis
}
\usage{
empirical_matlist
}
\description{
raw data for analysis of empirical data
}
\keyword{datasets}
