% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_data_phylo.R
\docType{data}
\name{phylo_key}
\alias{phylo_key}
\alias{phylo_data}
\alias{phylo_matrices}
\title{data for phylogenetic analysis of macaque hierarchy steepness}
\format{
\code{phylo_matrices} is a list of 67 interaction matrices and
\code{phylo_key} contains matching information on species and references
}
\source{
Christof Neumann
}
\usage{
phylo_key
}
\description{
data for phylogenetic analysis of macaque hierarchy steepness
}
\references{
see vignette for references
}
\keyword{datasets}
