% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_data_Primates301_nex.R
\docType{data}
\name{Primates301_nex}
\alias{Primates301_nex}
\title{consensus tree for primate species}
\format{
A object of class \code{phylo} (see \code{\link[ape]{phylo}})
}
\source{
\url{https://github.com/rmcelreath/rethinking/blob/master/data/Primates301_nex.rda}
}
\usage{
Primates301_nex
}
\description{
consensus tree for primate species
}
\references{
Arnold, C., Matthews, L.J. and Nunn, C.L. (2010), The 10kTrees website: A new online resource for primate phylogeny. Evol. Anthropol., 19: 114-118. \url{https://doi.org/10.1002/evan.20251} (\url{https://10ktrees.nunn-lab.org/})
}
\keyword{datasets}
