% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_data_randomization.R
\docType{data}
\name{randomization_matlist}
\alias{randomization_matlist}
\alias{randomization_numbers}
\alias{randomization_progress}
\title{raw data for analysis of empirical data}
\format{
\code{randomization_numbers} is a data frame with 500 observations,
which represent the results of the attempt to establish rough guides
as to how often a matrix needs to be translated into randomized
sequences.

\code{randomization_progress} is a data frame with 500 rows, which
was used to track progress in the analysis

\code{randomization_matlist} is a large list with the individual
interaction matrices that were used for the evaluation of
randomization numbers
}
\usage{
randomization_matlist
}
\description{
raw data for analysis of empirical data
}
\keyword{datasets}
