% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomized_sequence_from_matrix.R
\name{randomized_sequence_from_matrix}
\alias{randomized_sequence_from_matrix}
\title{convert interaction matrix to sequence}
\usage{
randomized_sequence_from_matrix(mat, n_rand = 1)
}
\arguments{
\item{mat}{square matrix}

\item{n_rand}{number of randomized sequences to produce (default is \code{1})}
}
\value{
a list with a matrix of winner ids and a matrix with loser ids
}
\description{
convert interaction matrix to sequence
}
\examples{
data(bonobos, package = "EloRating")
bonobos[1:4, 1:4]
randomized_sequence_from_matrix(bonobos[1:4, 1:4])
}
