% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slope_box.R
\name{slope_box}
\alias{slope_box}
\title{plot steepness as function of unknown relationships}
\usage{
slope_box(pdata, empirical = NULL, hinges = c(0.05, 0.95), ylims = c(-1.5, 1))
}
\arguments{
\item{pdata}{\code{data("removal_slopes")}}

\item{empirical}{logical, should empirical data be plotted (if \code{TRUE}),
or artificial data (if \code{FALSE}). Default is \code{NULL}, which
combines both sets.}

\item{hinges}{numeric of length 2, determines the quantiles of the boxplot.
Default is \code{c(0.05, 0.95)}.}

\item{ylims}{numeric of length 2, determines the y-limits.}
}
\value{
a plot, and invisibly the numeric data.
}
\description{
plot steepness as function of unknown relationships
}
\examples{
data("removal_slopes")
x <- slope_box(removal_slopes, empirical = TRUE)
}
