% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_data_toy_example.R
\docType{data}
\name{toy_example}
\alias{toy_example}
\title{two toy dominance networks}
\format{
A list.
\describe{
\item{m1,m2}{the two matrices}
\item{e1,e2}{results of \code{\link[EloSteepness]{elo_steepness_from_matrix}}}
\item{d1,d2}{results of \code{\link[EloSteepness]{davids_steepness}}}
}
}
\usage{
toy_example
}
\description{
Two toy examples with without stan_fit objects for Elo and DS.
}
\details{
For the sake of saving space, the actual stan_fit objects were
removed.
}
\keyword{datasets}
